Conclude the PR, added tests and checks
