+++ dotnet/src/Connectors/Connectors.Google.UnitTests/GeminiPromptExecutionSettingsTests.cs
@@ -2,6 +2,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text.Json;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.Connectors.Google;
@@ -109,7 +110,10 @@ public sealed class GeminiPromptExecutionSettingsTests
                               "category": "{{category.Label}}",
                               "threshold": "{{threshold.Label}}"
                             }
-                          ]
+                          ],
+                          "thinking_config": {
+                            "thinking_budget": 1000
+                          }
                         }
                         """;
         var actualSettings = JsonSerializer.Deserialize<PromptExecutionSettings>(json);
@@ -129,6 +133,8 @@ public sealed class GeminiPromptExecutionSettingsTests
         Assert.Single(executionSettings.SafetySettings!, settings =>
             settings.Category.Equals(category) &&
             settings.Threshold.Equals(threshold));
+
+        Assert.Equal(1000, executionSettings.ThinkingConfig?.ThinkingBudget);
     }
 
     [Fact]
@@ -152,7 +158,10 @@ public sealed class GeminiPromptExecutionSettingsTests
                               "category": "{{category.Label}}",
                               "threshold": "{{threshold.Label}}"
                             }
-                          ]
+                          ],
+                          "thinking_config": {
+                            "thinking_budget": 1000
+                          }
                         }
                         """;
         var executionSettings = JsonSerializer.Deserialize<GeminiPromptExecutionSettings>(json);
@@ -168,6 +177,7 @@ public sealed class GeminiPromptExecutionSettingsTests
         Assert.Equivalent(executionSettings.StopSequences, clone.StopSequences);
         Assert.Equivalent(executionSettings.SafetySettings, clone.SafetySettings);
         Assert.Equal(executionSettings.AudioTimestamp, clone.AudioTimestamp);
+        Assert.Equivalent(executionSettings.ThinkingConfig, clone.ThinkingConfig);
     }
 
     [Fact]
@@ -191,7 +201,10 @@ public sealed class GeminiPromptExecutionSettingsTests
                               "category": "{{category.Label}}",
                               "threshold": "{{threshold.Label}}"
                             }
-                          ]
+                          ],
+                          "thinking_config": {
+                            "thinking_budget": 1000
+                          }
                         }
                         """;
         var executionSettings = JsonSerializer.Deserialize<GeminiPromptExecutionSettings>(json);
@@ -206,5 +219,7 @@ public sealed class GeminiPromptExecutionSettingsTests
         Assert.Throws<InvalidOperationException>(() => executionSettings.Temperature = 0.5);
         Assert.Throws<InvalidOperationException>(() => executionSettings.AudioTimestamp = false);
         Assert.Throws<NotSupportedException>(() => executionSettings.StopSequences!.Add("baz"));
+        Assert.Throws<NotSupportedException>(() => executionSettings.SafetySettings!.Add(new GeminiSafetySetting(GeminiSafetyCategory.Toxicity, GeminiSafetyThreshold.Unspecified)));
+        Assert.Throws<InvalidOperationException>(() => executionSettings.ThinkingConfig = new GeminiThinkingConfig { ThinkingBudget = 1 });
     }
 }
