+++ dotnet/src/Connectors/Connectors.Google.UnitTests/Services/GoogleAIGeminiChatCompletionServiceTests.cs
@@ -4,6 +4,7 @@ using System;
 using System.IO;
 using System.Net.Http;
 using System.Text;
+using System.Text.Json;
 using System.Threading.Tasks;
 using Microsoft.SemanticKernel.ChatCompletion;
 using Microsoft.SemanticKernel.Connectors.Google;
@@ -99,7 +100,7 @@ public sealed class GoogleAIGeminiChatCompletionServiceTests : IDisposable
         if (shouldContain)
         {
             Assert.Contains("thinkingConfig", requestBody);
-            Assert.Contains("thinkingBudget", requestBody);
+            Assert.Contains($"\"thinkingBudget\":{thinkingBudget}", requestBody);
         }
         else
         {
